Backend/Geo_Cortex/Geo_Cortex_Assistant/requirements.txt
langchain
langchain-text-splitters
langchain-community
langchain-ollama
fastapi
uvicorn
pandas
sqlalchemy
python-dotenv
anyio
httpx
requests
faiss-cpu
psycopg2-binary
geoalchemy2
python-multipart
shapely
gpxpy
fastkml
